fixed bug regarding reseting state and chenged names of a couple components. #5
