--- client/components/chat-list-item.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function chatRoomListItem(props) {
-
-}
