+++ client/pages/home.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import ChatList from '../components/chat-list';
+import ChatListSection from '../components/chat-list-section';
 
 export default function Home(props) {
   return (
     <>
-      <ChatList />
+      <ChatListSection />
     </>
   );
 }
